Update Ipopt to v3.12.13

+++ Ipopt/Ipopt/src/Algorithm/LinearSolvers/IpMa57TSolverInterface.cpp
@@ -2,7 +2,7 @@
 // All Rights Reserved.
 // This code is published under the Eclipse Public License.
 //
-// $Id: IpMa57TSolverInterface.cpp 2613 2015-11-04 14:42:02Z stefan $
+// $Id$
 //
 // Authors:  Michael Hagemann               Univ of Basel 2005-10-28
 //               original version (based on MA27TSolverInterface.cpp)
@@ -654,7 +654,7 @@ namespace Ipopt
         Jnlst().Printf(J_WARNING, J_LINEAR_ALGEBRA,
                        "Reallocating memory for MA57: lfact (%d)\n", wd_lfact_);
 
-        if( wd_lfact_ > std::numeric_limits<size_t>::max() / sizeof(double) ) {
+        if( (size_t)wd_lfact_ > std::numeric_limits<size_t>::max() / sizeof(double) ) {
           Jnlst().Printf(J_ERROR, J_LINEAR_ALGEBRA,
                        "Cannot allocate memory of size %d exceeding SIZE_MAX = %u\n", wd_lfact_, std::numeric_limits<size_t>::max());
           return SYMSOLVER_FATAL_ERROR;
